refactor: enums package location
