+++ src/main/java/com/pdg/pymesbackend/model/Question.java
@@ -1,6 +1,6 @@
 package com.pdg.pymesbackend.model;
 
-import com.pdg.pymesbackend.model.enums.QuestionType;
+import com.pdg.pymesbackend.enums.QuestionType;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
